Student flow with one task
