+++ MyApp/app/(tabs)/h1.tsx
@@ -1,196 +1,131 @@
 import { View, Text, StyleSheet, FlatList, SafeAreaView, TouchableOpacity } from 'react-native'
-import React, { useState, useEffect } from 'react'
-import { useRouter, useLocalSearchParams } from 'expo-router'
+import React, { useState } from 'react'
+import { useRouter } from 'expo-router'
+import { Ionicons } from '@expo/vector-icons'
 
-interface Tehtava {
+interface KurssiTehtava {
   id: string
   nimi: string
-  status: 'not_started' | 'submitted' | 'approved'
-  suoritettuPvm?: string
-  itsearviointi?: string
+  progress?: number // Progress percentage for visual indicator
 }
 
-interface Kurssi {
+interface Oppiaine {
   id: string
   nimi: string
-  tehtavat: Tehtava[]
+  tehtavat: KurssiTehtava[]
+  expanded: boolean
 }
 
-const StudentView = () => {
+const CourseSelectionView = () => {
   const router = useRouter()
-  const params = useLocalSearchParams()
-  const [kurssit, setKurssit] = useState<Kurssi[]>([
+  const [oppiaineet, setOppiaineet] = useState<Oppiaine[]>([
     {
       id: '1',
-      nimi: 'H1 Syksy',
+      nimi: 'Kariologia',
+      expanded: false,
       tehtavat: [
-        { id: '1', nimi: 'Hampaiden tunnistus', status: 'approved', suoritettuPvm: '29.9.2025', itsearviointi: 'Tehtävä sujui hyvin, opin tunnistamaan eri hampaat.' },
-        { id: '2', nimi: 'Käsi-instrumentteihin tutustuminen', status: 'not_started' },
-      ],
+        { id: '1', nimi: 'H1 Syksy', progress: 50 },
+        { id: '2', nimi: 'H1 Kevät', progress: 0 },
+        { id: '3', nimi: 'H2 Syksy', progress: 0 },
+        { id: '4', nimi: 'H2 Kevät', progress: 0 },
+        { id: '5', nimi: 'H3 Syksy', progress: 0 },
+        { id: '6', nimi: 'H3 Kevät', progress: 0 },
+        { id: '7', nimi: 'Mini-OSCE', progress: 0 },
+      ]
     },
+    {
+      id: '2',
+      nimi: 'Kirurgia',
+      expanded: false,
+      tehtavat: [
+        { id: '8', nimi: 'H3 Aseptiikan ryhmäopetus', progress: 0 },
+        { id: '9', nimi: 'H3 Puudutus-harjoitus', progress: 0 },
+        { id: '10', nimi: 'H3 Hampaan poistoharjoitus', progress: 0 },
+        { id: '11', nimi: 'H3 Ompelu-harjoitus', progress: 0 },
+        { id: '12', nimi: 'H4 Leikkaus-harjoitus', progress: 0 },
+      ]
+    }
   ])
 
-  // Handle updated task data from task detail page
-  useEffect(() => {
-    if (params.updatedTask) {
-      try {
-        const updatedTaskData = JSON.parse(params.updatedTask as string)
-
-        // Find and update the matching task using task ID
-        setKurssit(prev =>
-          prev.map(kurssi => ({
-            ...kurssi,
-            tehtavat: kurssi.tehtavat.map(task =>
-              task.id === updatedTaskData.id
-                ? {
-                  ...task,
-                  status: updatedTaskData.status,
-                  suoritettuPvm: updatedTaskData.completionDate,
-                  itsearviointi: updatedTaskData.selfAssessment
-                }
-                : task
-            )
-          }))
-        )
-
-        // Clear the params to avoid re-processing
-        router.setParams({ updatedTask: undefined })
-      } catch (error) {
-        console.error('Error parsing updated task data:', error)
-      }
-    }
-  }, [params.updatedTask])
-
-  const updateTaskStatus = (kurssiId: string, tehtavaId: string, newStatus: 'not_started' | 'submitted' | 'approved', completionDate?: string, selfAssessment?: string) => {
-    setKurssit(prev =>
-      prev.map(kurssi =>
-        kurssi.id === kurssiId
-          ? {
-            ...kurssi,
-            tehtavat: kurssi.tehtavat.map(t =>
-              t.id === tehtavaId
-                ? {
-                  ...t,
-                  status: newStatus,
-                  suoritettuPvm: completionDate || t.suoritettuPvm,
-                  itsearviointi: selfAssessment || t.itsearviointi
-                }
-                : t
-            ),
-          }
-          : kurssi
+  const toggleOppiaine = (oppiaineId: string) => {
+    setOppiaineet(prev =>
+      prev.map(oppiaine =>
+        oppiaine.id === oppiaineId
+          ? { ...oppiaine, expanded: !oppiaine.expanded }
+          : oppiaine
       )
     )
   }
 
-  const handleTaskPress = (task: Tehtava) => {
-    router.push({
-      pathname: '/task-detail',
-      params: {
-        taskId: task.id,
-        name: task.nimi,
-        status: task.status,
-        completionDate: task.suoritettuPvm || '',
-        selfAssessment: task.itsearviointi || ''
-      }
-    })
-  }
-
-  const handleStatusButtonPress = (kurssiId: string, task: Tehtava) => {
-    if (task.status === 'not_started') {
-      // Navigate to task detail page for new task
-      handleTaskPress(task)
+  const handleCoursePress = (tehtava: KurssiTehtava) => {
+    // Navigate to specific course page (e.g., H1 tasks)
+    if (tehtava.nimi === 'H1 Syksy') {
+      router.push({
+        pathname: '/h1-tasks' as any // TypeScript workaround for dynamic routes
+      })
     }
-    // For submitted/approved tasks, clicking the status doesn't do anything
-    // They should click on the task name to view details
+    // Add more specific routes for other courses as needed
   }
 
-  const getStatusText = (status: string) => {
-    switch (status) {
-      case 'not_started':
-        return 'Tee'
-      case 'submitted':
-        return 'Odottaa hyväksyntää'
-      case 'approved':
-        return 'Hyväksytty'
-      default:
-        return 'Tee'
-    }
-  }
-
-  const getStatusButtonStyle = (status: string) => {
-    switch (status) {
-      case 'not_started':
-        return [styles.statusButton]
-      case 'submitted':
-        return [styles.statusButton, styles.statusButtonSubmitted]
-      case 'approved':
-        return [styles.statusButton, styles.statusButtonApproved]
-      default:
-        return [styles.statusButton]
-    }
-  }
-
-  const getStatusTextStyle = (status: string) => {
-    switch (status) {
-      case 'not_started':
-        return [styles.statusText]
-      case 'submitted':
-        return [styles.statusText, styles.statusTextSubmitted]
-      case 'approved':
-        return [styles.statusText, styles.statusTextApproved]
-      default:
-        return [styles.statusText]
-    }
-  }
-
-  const renderTehtava = (kurssiId: string) => ({ item }: { item: Tehtava }) => (
-    <View style={styles.tehtavaCard}>
-      <TouchableOpacity
-        style={{ flex: 1 }}
-        onPress={() => handleTaskPress(item)}
-      >
-        <Text style={[styles.tehtavaNimi, item.status !== 'not_started' && styles.tehtavaNimiClickable]}>{item.nimi}</Text>
-        {item.status !== 'not_started' && item.suoritettuPvm && (
-          <Text style={styles.suoritettu}>Suoritettu: {item.suoritettuPvm}</Text>
-        )}
-      </TouchableOpacity>
-
-      <TouchableOpacity
-        onPress={() => handleStatusButtonPress(kurssiId, item)}
-        style={getStatusButtonStyle(item.status)}
-        disabled={item.status !== 'not_started'}
-      >
-        <Text style={getStatusTextStyle(item.status)}>
-          {getStatusText(item.status)}
-        </Text>
-      </TouchableOpacity>
+  const renderProgressBar = (progress: number) => (
+    <View style={styles.progressContainer}>
+      <View style={styles.progressBar}>
+        <View style={[styles.progressFill, { width: `${progress}%` }]} />
+      </View>
     </View>
   )
 
-  const renderKurssi = ({ item }: { item: Kurssi }) => (
-    <View style={styles.kurssiCard}>
-      <Text style={styles.kurssiTitle}>{item.nimi}</Text>
-      <FlatList
-        data={item.tehtavat}
-        renderItem={renderTehtava(item.id)}
-        keyExtractor={t => t.id}
-      />
+  const renderTehtava = ({ item }: { item: KurssiTehtava }) => (
+    <TouchableOpacity
+      style={styles.tehtavaItem}
+      onPress={() => handleCoursePress(item)}
+    >
+      <Text style={styles.tehtavaNimi}>{item.nimi}</Text>
+      {renderProgressBar(item.progress || 0)}
+    </TouchableOpacity>
+  )
+
+  const renderOppiaine = ({ item }: { item: Oppiaine }) => (
+    <View style={styles.oppiaineCard}>
+      <TouchableOpacity
+        style={styles.oppiaineHeader}
+        onPress={() => toggleOppiaine(item.id)}
+      >
+        <Text style={styles.oppiaineTitle}>{item.nimi}</Text>
+        <Ionicons
+          name={item.expanded ? "chevron-up" : "chevron-down"}
+          size={24}
+          color="#333"
+        />
+      </TouchableOpacity>
+
+      {item.expanded && (
+        <View style={styles.tehtavaList}>
+          <FlatList
+            data={item.tehtavat}
+            renderItem={renderTehtava}
+            keyExtractor={t => t.id}
+            scrollEnabled={false}
+          />
+        </View>
+      )}
     </View>
   )
 
   return (
     <SafeAreaView style={styles.container}>
+      <Text style={styles.headerTitle}>Suoritta harjoituksia!</Text>
       <FlatList
-        data={kurssit}
-        renderItem={renderKurssi}
-        keyExtractor={k => k.id}
+        data={oppiaineet}
+        renderItem={renderOppiaine}
+        keyExtractor={item => item.id}
+        showsVerticalScrollIndicator={false}
       />
     </SafeAreaView>
   )
 }
 
-export default StudentView
+export default CourseSelectionView
 
 const styles = StyleSheet.create({
   container: {
@@ -198,82 +133,60 @@ const styles = StyleSheet.create({
     backgroundColor: '#f5f5f5',
     padding: 16,
   },
-  kurssiCard: {
-    backgroundColor: '#fff',
-    borderRadius: 10,
-    padding: 16,
+  headerTitle: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    color: '#333',
+    textAlign: 'center',
     marginBottom: 20,
+  },
+  oppiaineCard: {
+    backgroundColor: '#fff',
+    borderRadius: 12,
+    marginBottom: 16,
     shadowColor: '#000',
-    shadowOpacity: 0.05,
+    shadowOpacity: 0.1,
     shadowRadius: 4,
     shadowOffset: { width: 0, height: 2 },
-    elevation: 2,
+    elevation: 3,
   },
-  kurssiTitle: {
+  oppiaineHeader: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    padding: 16,
+  },
+  oppiaineTitle: {
     fontSize: 20,
     fontWeight: '600',
-    color: '#222',
-    borderBottomWidth: 1,
-    borderBottomColor: '#ddd',
-    paddingBottom: 8,
-    marginBottom: 12,
-    textAlign: 'center',
+    color: '#333',
   },
-  tehtavaCard: {
-    backgroundColor: '#fafafa',
-    borderRadius: 8,
-    padding: 12,
-    marginBottom: 10,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    borderWidth: 1,
-    borderColor: '#e0e0e0',
+  tehtavaList: {
+    paddingBottom: 8,
+  },
+  tehtavaItem: {
+    padding: 16,
+    borderTopWidth: 1,
+    borderTopColor: '#f0f0f0',
   },
   tehtavaNimi: {
     fontSize: 16,
     fontWeight: '500',
     color: '#333',
-    marginBottom: 4,
+    marginBottom: 8,
   },
-  tehtavaNimiClickable: {
-    color: '#007AFF',
-    textDecorationLine: 'underline',
+  progressContainer: {
+    marginTop: 4,
   },
-  suoritettu: {
-    fontSize: 13,
-    color: '#555',
-  },
-  hyvaksyntatila: {
-    fontSize: 13,
-    color: '#888',
-  },
-  statusButton: {
+  progressBar: {
+    height: 6,
     backgroundColor: '#e0e0e0',
-    borderRadius: 20,
-    paddingVertical: 6,
-    paddingHorizontal: 14,
+    borderRadius: 3,
+    overflow: 'hidden',
   },
-  statusButtonDone: {
-    backgroundColor: '#333',
-  },
-  statusButtonSubmitted: {
-    backgroundColor: '#FFA500',
-  },
-  statusButtonApproved: {
-    backgroundColor: '#4CAF50',
-  },
-  statusText: {
-    color: '#333',
-    fontWeight: '600',
-  },
-  statusTextDone: {
-    color: '#fff',
-  },
-  statusTextSubmitted: {
-    color: '#fff',
-  },
-  statusTextApproved: {
-    color: '#fff',
+  progressFill: {
+    height: '100%',
+    backgroundColor: '#007AFF',
+    borderRadius: 3,
   },
 })
