+++ MyApp/app/(tabs)/student.tsx
@@ -1,105 +1,125 @@
 import { View, Text, StyleSheet, FlatList, SafeAreaView, TouchableOpacity } from 'react-native'
 import React, { useState } from 'react'
+import { useRouter } from 'expo-router'
+import { Ionicons } from '@expo/vector-icons'
 
-interface Tehtava {
+interface KurssiTehtava {
   id: string
   nimi: string
-  tehty: boolean
+  progress?: number // Progress percentage for visual indicator
 }
 
-interface Kurssi {
+interface Oppiaine {
   id: string
   nimi: string
-  tehtavat: Tehtava[]
+  tehtavat: KurssiTehtava[]
+  expanded: boolean
 }
 
 const StudentView = () => {
-  const [kurssit, setKurssit] = useState<Kurssi[]>([
+  const router = useRouter()
+  const [oppiaineet, setOppiaineet] = useState<Oppiaine[]>([
     {
       id: '1',
       nimi: 'Kariologia',
+      expanded: false,
       tehtavat: [
-        { id: '1', nimi: 'H1 Syksy', tehty: true },
-        { id: '2', nimi: 'H1 Kevät', tehty: false },
-        { id: '3', nimi: 'H2 Syksy', tehty: false },
-        { id: '4', nimi: 'H2 Kevät', tehty: false },
-        { id: '5', nimi: 'H3 Syksy', tehty: false },
-        { id: '6', nimi: 'H3 Kevät', tehty: false },
-        { id: '7', nimi: 'Mini-OSCE', tehty: false },
-      ],
+        { id: '1', nimi: 'H1 Syksy', progress: 50 },
+        { id: '2', nimi: 'H1 Kevät', progress: 0 },
+        { id: '3', nimi: 'H2 Syksy', progress: 0 },
+        { id: '4', nimi: 'H2 Kevät', progress: 0 },
+        { id: '5', nimi: 'H3 Syksy', progress: 0 },
+        { id: '6', nimi: 'H3 Kevät', progress: 0 },
+        { id: '7', nimi: 'Mini-OSCE', progress: 0 },
+      ]
     },
     {
       id: '2',
       nimi: 'Kirurgia',
+      expanded: false,
       tehtavat: [
-        { id: '1', nimi: 'H3 Aseptiikan ryhmäopetus', tehty: false },
-        { id: '2', nimi: 'H3 Puudutus-harjoitus', tehty: false },
-        { id: '3', nimi: 'H3 Hampaan poistoharjoitus', tehty: false },
-        { id: '4', nimi: 'H3 Ompeluharjoitus', tehty: false },
-        { id: '5', nimi: 'H4 Leikkausharjoitus', tehty: false },
-      ],
-    },
+        { id: '8', nimi: 'H3 Aseptiikan ryhmäopetus', progress: 0 },
+        { id: '9', nimi: 'H3 Puudutus-harjoitus', progress: 0 },
+        { id: '10', nimi: 'H3 Hampaan poistoharjoitus', progress: 0 },
+        { id: '11', nimi: 'H3 Ompelu-harjoitus', progress: 0 },
+        { id: '12', nimi: 'H4 Leikkaus-harjoitus', progress: 0 },
+      ]
+    }
   ])
 
-  const [openKurssiId, setOpenKurssiId] = useState<string | null>(null)
-
-  const toggleTehtava = (kurssiId: string, tehtavaId: string) => {
-    setKurssit(prev =>
-      prev.map(kurssi =>
-        kurssi.id === kurssiId
-          ? {
-              ...kurssi,
-              tehtavat: kurssi.tehtavat.map(tehtava =>
-                tehtava.id === tehtavaId
-                  ? { ...tehtava, tehty: !tehtava.tehty }
-                  : tehtava
-              ),
-            }
-          : kurssi
+  const toggleOppiaine = (oppiaineId: string) => {
+    setOppiaineet(prev =>
+      prev.map(oppiaine =>
+        oppiaine.id === oppiaineId
+          ? { ...oppiaine, expanded: !oppiaine.expanded }
+          : oppiaine
       )
     )
   }
 
-  const renderTehtava = (kurssiId: string) => ({ item }: { item: Tehtava }) => {
-    const progressWidth = item.tehty ? '80%' : '10%'
-    return (
-      <TouchableOpacity onPress={() => toggleTehtava(kurssiId, item.id)}>
-        <View style={styles.tehtavaCard}>
-          <Text style={styles.tehtavaText}>{item.nimi}</Text>
-          <View style={styles.progressBarBackground}>
-            <View style={[styles.progressBarFill, { width: progressWidth }]} />
-          </View>
-        </View>
-      </TouchableOpacity>
-    )
+  const handleCoursePress = (tehtava: KurssiTehtava) => {
+    // Navigate to specific course page (e.g., H1 tasks)
+    if (tehtava.nimi === 'H1 Syksy') {
+      router.push({
+        pathname: '/h1-tasks' as any // TypeScript workaround for dynamic routes
+      })
+    }
+    // Add more specific routes for other courses as needed
   }
 
-  const renderKurssi = ({ item }: { item: Kurssi }) => {
-    const isOpen = openKurssiId === item.id
-    return (
-      <View style={styles.kurssiCard}>
-        <TouchableOpacity onPress={() => setOpenKurssiId(isOpen ? null : item.id)} style={styles.kurssiHeader}>
-          <Text style={styles.kurssiTitle}>{item.nimi}</Text>
-          <Text style={styles.arrow}>{isOpen ? '▲' : '▼'}</Text>
-        </TouchableOpacity>
+  const renderProgressBar = (progress: number) => (
+    <View style={styles.progressContainer}>
+      <View style={styles.progressBar}>
+        <View style={[styles.progressFill, { width: `${progress}%` }]} />
+      </View>
+    </View>
+  )
 
-        {isOpen && (
+  const renderTehtava = ({ item }: { item: KurssiTehtava }) => (
+    <TouchableOpacity
+      style={styles.tehtavaItem}
+      onPress={() => handleCoursePress(item)}
+    >
+      <Text style={styles.tehtavaNimi}>{item.nimi}</Text>
+      {renderProgressBar(item.progress || 0)}
+    </TouchableOpacity>
+  )
+
+  const renderOppiaine = ({ item }: { item: Oppiaine }) => (
+    <View style={styles.oppiaineCard}>
+      <TouchableOpacity
+        style={styles.oppiaineHeader}
+        onPress={() => toggleOppiaine(item.id)}
+      >
+        <Text style={styles.oppiaineTitle}>{item.nimi}</Text>
+        <Ionicons
+          name={item.expanded ? "chevron-up" : "chevron-down"}
+          size={24}
+          color="#333"
+        />
+      </TouchableOpacity>
+
+      {item.expanded && (
+        <View style={styles.tehtavaList}>
           <FlatList
             data={item.tehtavat}
-            renderItem={renderTehtava(item.id)}
+            renderItem={renderTehtava}
             keyExtractor={t => t.id}
+            scrollEnabled={false}
           />
-        )}
-      </View>
-    )
-  }
+        </View>
+      )}
+    </View>
+  )
 
   return (
     <SafeAreaView style={styles.container}>
+      <Text style={styles.headerTitle}>Tervetuloa Taitopaja Appiin!</Text>
       <FlatList
-        data={kurssit}
-        renderItem={renderKurssi}
-        keyExtractor={k => k.id}
+        data={oppiaineet}
+        renderItem={renderOppiaine}
+        keyExtractor={item => item.id}
+        showsVerticalScrollIndicator={false}
       />
     </SafeAreaView>
   )
@@ -110,47 +130,63 @@ export default StudentView
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    padding: 20,
-    backgroundColor: '#f0f0f0',
+    backgroundColor: '#f5f5f5',
+    padding: 16,
   },
-  kurssiCard: {
+  headerTitle: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    color: '#333',
+    textAlign: 'center',
+    marginBottom: 20,
+  },
+  oppiaineCard: {
     backgroundColor: '#fff',
     borderRadius: 12,
-    marginBottom: 15,
-    padding: 10,
+    marginBottom: 16,
     shadowColor: '#000',
     shadowOpacity: 0.1,
+    shadowRadius: 4,
     shadowOffset: { width: 0, height: 2 },
-    shadowRadius: 5,
+    elevation: 3,
   },
-  kurssiHeader: {
+  oppiaineHeader: {
     flexDirection: 'row',
     justifyContent: 'space-between',
     alignItems: 'center',
+    padding: 16,
   },
-  kurssiTitle: {
+  oppiaineTitle: {
     fontSize: 20,
-    fontWeight: 'bold',
+    fontWeight: '600',
+    color: '#333',
   },
-  arrow: {
-    fontSize: 20,
+  tehtavaList: {
+    paddingBottom: 8,
   },
-  tehtavaCard: {
-    marginTop: 10,
+  tehtavaItem: {
+    padding: 16,
+    borderTopWidth: 1,
+    borderTopColor: '#f0f0f0',
   },
-  tehtavaText: {
+  tehtavaNimi: {
     fontSize: 16,
-    marginBottom: 5,
+    fontWeight: '500',
+    color: '#333',
+    marginBottom: 8,
   },
-  progressBarBackground: {
-    width: '100%',
+  progressContainer: {
+    marginTop: 4,
+  },
+  progressBar: {
     height: 6,
     backgroundColor: '#e0e0e0',
     borderRadius: 3,
+    overflow: 'hidden',
   },
-  progressBarFill: {
-    height: 6,
-    backgroundColor: '#007bff',
+  progressFill: {
+    height: '100%',
+    backgroundColor: '#007AFF',
     borderRadius: 3,
   },
 })
